Move settings into its own subdirectory
