+++ src/server.ts
@@ -9,14 +9,9 @@ import bodyParser from "body-parser";
 import { authenticateCookie, enforceLoggedIn } from "./auth/middlewares/authenticate";
 import { api as unauthenticatedApi, views as unauthenticatedViews } from "./toolbox/routers/unauthenticated";
 import { views as secretDjViews } from "./secret-dj/routers/views";
-import subdomain from "express-subdomain";
 import { adminViews } from "./toolbox/routers/views";
-import emails from "./toolbox/storage/emails";
-import { parse } from "./parse";
-import { match, P } from "ts-pattern";
-import { Ok } from "./types/result";
+import { views as settingsViews } from "./settings/routers/views";
 import session from "express-session";
-import { settingsSchema } from "./schema";
 import { gateKeepInvalidURIs } from "./common/middlewares/gateKeepInvalidURIs";
 
 // TODO would be great to also serve a javascript client
@@ -35,30 +30,7 @@ const views = express()
 	.use(unauthenticatedViews)
 	.use("/toolbox", enforceLoggedIn, adminViews)
 	.use(enforceLoggedIn)
-	.get("/settings", (req, res) =>
-		res.render("pages/settings", {
-			subscribed: res.locals.email.subscribed,
-			error: "",
-			success: "",
-			Config,
-		}),
-	)
-	.post("/settings", async (req, res) => {
-		const result = parse(settingsSchema, req.body);
-		return match(result)
-			.with(Ok(P.select()), async ({ subscribed }) => {
-				await emails.updateSettings(res.locals.email.id, subscribed);
-				return res.render("pages/settings", { subscribed, error: "", success: "saved", Config });
-			})
-			.otherwise(() =>
-				res.render("pages/settings", {
-					subscribed: res.locals.email.subscribed,
-					error: "that didn't work",
-					success: "",
-					Config,
-				}),
-			);
-	})
+	.use("/settings", settingsViews)
 	.get("/", (req, res) => res.render("pages/index"))
 	.get("/*", (req, res) => res.sendStatus(404));
